Add remote zipped toolkit context
