+++ java/src/com/ibm/streamsx/topology/internal/context/ZippedToolkitStreamsContext.java
@@ -1,27 +1,18 @@
 package com.ibm.streamsx.topology.internal.context;
 
+import static com.ibm.streamsx.topology.context.ContextProperties.KEEP_ARTIFACTS;
+import static com.ibm.streamsx.topology.internal.json4j.JSON4JUtilities.gson;
+
 import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.PrintWriter;
-import java.net.URISyntaxException;
-import java.nio.file.FileVisitResult;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Map;
 import java.util.concurrent.Future;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
 
 import com.ibm.json.java.JSONObject;
-import com.ibm.json.java.OrderedJSONObject;
 import com.ibm.streamsx.topology.Topology;
-import com.ibm.streamsx.topology.internal.process.CompletedFuture;
+import com.ibm.streamsx.topology.context.ContextProperties;
+import com.ibm.streamsx.topology.context.remote.RemoteContext;
+import com.ibm.streamsx.topology.context.remote.RemoteContextFactory;
+import com.ibm.streamsx.topology.internal.context.remote.ZippedToolkitRemoteContext;
 
 public class ZippedToolkitStreamsContext extends ToolkitStreamsContext {
 	
@@ -34,105 +25,25 @@ public class ZippedToolkitStreamsContext extends ToolkitStreamsContext {
 	public Future<File> submit(Topology app, Map<String, Object> config) throws Exception {        
         File toolkitRoot = super.submit(app, config).get();
         
-        String namespace = (String) app.builder().json().get("namespace");
-        String name = (String) app.builder().json().get("name");
-        String tkName = toolkitRoot.toPath().toAbsolutePath().getFileName().toString();
-        
-        Path zipOutPath = pack(toolkitRoot.toPath(), namespace, name, tkName);
-        
-        JSONObject jso = new JSONObject();
-        addConfigToJSON(jso, config);  
-        
-        return deleteToolkitAndProduceCompletedFuture(toolkitRoot, jso, zipOutPath);
+        JSONObject deploy = new JSONObject();
+        deploy.put(ContextProperties.TOOLKIT_DIR, toolkitRoot.getAbsolutePath());
+        if (config.containsKey(ContextProperties.KEEP_ARTIFACTS))
+            deploy.put(KEEP_ARTIFACTS, config.get(KEEP_ARTIFACTS));
+
+        JSONObject submission = new JSONObject();
+        submission.put(SUBMISSION_DEPLOY, deploy);
+        submission.put(SUBMISSION_GRAPH, app.builder().complete());
+        return ZippedToolkitRemoteContext.createCodeArchive(toolkitRoot, gson(submission));
 	}
 	
 	@Override
 	public Future<File> submit(JSONObject submission) throws Exception {
-        File toolkitRoot = super.submit(submission).get();
-        
-        JSONObject jsonGraph = (JSONObject) submission.get(SUBMISSION_GRAPH);
-        String namespace = jsonGraph.get("namespace").toString();
-        String name = jsonGraph.get("name").toString();
-        String tkName = toolkitRoot.toPath().toAbsolutePath().getFileName().toString();
-        
-        Path zipOutPath = pack(toolkitRoot.toPath(), namespace, name, tkName);
-        
-        return deleteToolkitAndProduceCompletedFuture(toolkitRoot, (JSONObject)submission.get(SUBMISSION_DEPLOY), zipOutPath);
+	    // Let the remote archive do all the work.
+	    @SuppressWarnings("unchecked")
+        RemoteContext<File> ztrc = (RemoteContext<File>) RemoteContextFactory.getRemoteContext(RemoteContext.Type.ZIPPED_TOOLKIT);
+	    
+	    return ztrc.submit(gson(submission));
 	}
-	
-	// Addresses 
-	private CompletedFuture<File> deleteToolkitAndProduceCompletedFuture(File toolkitRoot, JSONObject deploy, Path zipOutPath) throws IOException{
-		deleteToolkit(toolkitRoot, deploy);
-        return new CompletedFuture<File>(zipOutPath.toFile());
-	}
-	
-	public Path pack(final Path folder, String namespace, String name, String tkName) throws IOException, URISyntaxException {
-		Path zipFilePath = Paths.get(folder.toAbsolutePath().toString() + ".zip");
-		String workingDir = zipFilePath.getParent().toString();
-		// com.ibm.streamsx.topology/lib/com.ibm.streamsx.topology.jar
-		File jarLocation = new File(ZippedToolkitStreamsContext.class.getProtectionDomain().getCodeSource().getLocation().toURI().getPath());
-		// com.ibm.streamsx.topology/lib/
-		File libDir = new File(jarLocation.getParent());
-		// com.ibm.streamsx.topology
-		File topologyToolkit = new File(libDir.getParent());	
-		
-		// tkManifest is the list of toolkits contained in the archive
-		PrintWriter tkManifest = new PrintWriter("manifest_tk.txt", "UTF-8");
-		
-		// mainComposite is a string of the namespace and the main composite.
-		// This is used by the Makefile
-		PrintWriter mainComposite = new PrintWriter("main_composite.txt", "UTF-8");
-		
-		tkManifest.println(tkName);
-		tkManifest.println("com.ibm.streamsx.topology");
-		
-		mainComposite.print(namespace + "::" + name);
-		
-		tkManifest.close();
-		mainComposite.close();
-		
-		List<Path> paths = new ArrayList<Path>();
-		paths.add(Paths.get(topologyToolkit.getAbsolutePath()));
-		paths.add(Paths.get(workingDir + "/manifest_tk.txt"));
-		paths.add(Paths.get(workingDir + "/main_composite.txt"));
-		paths.add(Paths.get(topologyToolkit.getAbsolutePath() + "/opt/python/templates/common/Makefile"));
-		paths.add(folder);
-		
-		addAllToZippedArchive(paths, zipFilePath);	
-		paths.get(1).toFile().delete();
-		paths.get(2).toFile().delete();
-		
-	    return zipFilePath;
-	}
-	
-	protected static void addAllToZippedArchive(List<Path> starts, Path zipFilePath) throws IOException {
-		try (FileOutputStream fos = new FileOutputStream(zipFilePath.toFile());
-				ZipOutputStream zos = new ZipOutputStream(fos)) {
-			for (Path start : starts) {
-				String startName = start.getFileName().toString();
-				Files.walkFileTree(start, new SimpleFileVisitor<Path>() {
-					public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-						String entryName = startName;
-						String relativePath = start.relativize(file).toString();
-						if(relativePath.length() != 0){
-							entryName = entryName + "/" + relativePath;
-						}
-						zos.putNextEntry(new ZipEntry(entryName));
-						Files.copy(file, zos);
-						zos.closeEntry();
-						return FileVisitResult.CONTINUE;
-					}
-
-					public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-						zos.putNextEntry(new ZipEntry(startName + "/" + start.relativize(dir).toString() + "/"));
-						zos.closeEntry();
-						return FileVisitResult.CONTINUE;
-					}
-				});
-			}
-		}
-	}
-	
 	
 	@Override
     protected void makeToolkit(JSONObject deployInfo, File toolkitRoot) throws InterruptedException, Exception{
