+++ java/src/com/ibm/streamsx/topology/internal/file/FileUtilities.java
@@ -1,0 +1,48 @@
+package com.ibm.streamsx.topology.internal.file;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.FileVisitResult;
+import java.nio.file.FileVisitor;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.BasicFileAttributes;
+
+public class FileUtilities {
+    
+    /**
+     * Delete a directory and all of its files.
+     */
+    public static void deleteDirectory(File dir) throws IOException {
+
+        Files.walkFileTree(dir.toPath(), new FileVisitor<Path>() {
+
+            @Override
+            public FileVisitResult preVisitDirectory(Path dir,
+                    BasicFileAttributes attrs) throws IOException {
+                return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult visitFile(Path file,
+                    BasicFileAttributes attrs) throws IOException {
+                file.toFile().delete();
+                return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult visitFileFailed(Path file, IOException exc)
+                    throws IOException {
+                return FileVisitResult.CONTINUE;
+            }
+
+            @Override
+            public FileVisitResult postVisitDirectory(Path dir, IOException exc)
+                    throws IOException {
+                dir.toFile().delete();
+                return FileVisitResult.CONTINUE;
+            }
+        });
+    }
+
+}
